+++ authexample/src/main/java/com/example/authexample/HomeActivity.java
@@ -12,7 +12,7 @@ import android.widget.ArrayAdapter;
 import android.widget.ListView;
 import android.widget.TextView;
 
-import com.sourcepoint.cmplibrary.Consent;
+import com.sourcepoint.ccpa_cmplibrary.Consent;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashSet;
