+++ ccpa_cmplibrary/src/main/java/com/sourcepoint/ccpa_cmplibrary/Consent.java
@@ -1,4 +1,4 @@
-package com.sourcepoint.cmplibrary;
+package com.sourcepoint.ccpa_cmplibrary;
 
 import org.json.JSONException;
 import org.json.JSONObject;
