+++ ccpa_cmplibrary/src/main/java/com/sourcepoint/ccpa_cmplibrary/ConsentWebView.java
@@ -1,4 +1,4 @@
-package com.sourcepoint.cmplibrary;
+package com.sourcepoint.ccpa_cmplibrary;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
