+++ ccpa_cmplibrary/src/main/java/com/sourcepoint/ccpa_cmplibrary/SourcePointClient.java
@@ -1,4 +1,4 @@
-package com.sourcepoint.cmplibrary;
+package com.sourcepoint.ccpa_cmplibrary;
 
 import android.text.TextUtils;
 import android.util.Log;
@@ -12,8 +12,6 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.UnsupportedEncodingException;
-import java.net.URL;
-import java.net.URLEncoder;
 import java.util.HashSet;
 import java.util.UUID;
