+++ ccpa_cmplibrary/src/main/java/com/sourcepoint/ccpa_cmplibrary/UserConsent.java
@@ -1,4 +1,4 @@
-package com.sourcepoint.cmplibrary;
+package com.sourcepoint.ccpa_cmplibrary;
 
 import org.json.JSONArray;
 import org.json.JSONException;
