+++ ccpa_cmplibrary/src/test/java/com/sourcepoint/ccpa_cmplibrary/CCPAConsentLibBuilderTest.java
@@ -1,4 +1,4 @@
-package com.sourcepoint.cmplibrary;
+package com.sourcepoint.ccpa_cmplibrary;
 
 import android.app.Activity;
 import android.view.ViewGroup;
