+++ ccpa_cmplibrary/src/test/java/com/sourcepoint/ccpa_cmplibrary/EncodedParamTest.java
@@ -1,4 +1,4 @@
-package com.sourcepoint.cmplibrary;
+package com.sourcepoint.ccpa_cmplibrary;
 
 import org.junit.Test;
